Added the GL_Draw.h and .cpp to hold the OpenGL drawing functionality and to further abstract away actual OpenGL code drom the user code. Also did a little clean-up of the main source code.
