+++ OpenGL_Shaders_Example/Spinning_Cube/Debug/Spinning_Cube.tlog/Cl.items.tlog
@@ -1,3 +1,4 @@
+C:\Users\11307533\Documents\Programming\C++\OpenGL_Projects\Cpp_OpenGL_SpinningCube\OpenGL_Shaders_Example\Spinning_Cube\GL_Draw.cpp;C:\Users\11307533\Documents\Programming\C++\OpenGL_Projects\Cpp_OpenGL_SpinningCube\OpenGL_Shaders_Example\Spinning_Cube\Debug\GL_Draw.obj
 C:\Users\11307533\Documents\Programming\C++\OpenGL_Projects\Cpp_OpenGL_SpinningCube\OpenGL_Shaders_Example\Spinning_Cube\GL_ErrorHandeling.cpp;C:\Users\11307533\Documents\Programming\C++\OpenGL_Projects\Cpp_OpenGL_SpinningCube\OpenGL_Shaders_Example\Spinning_Cube\Debug\GL_ErrorHandeling.obj
 C:\Users\11307533\Documents\Programming\C++\OpenGL_Projects\Cpp_OpenGL_SpinningCube\OpenGL_Shaders_Example\Spinning_Cube\GL_ShaderProgram.cpp;C:\Users\11307533\Documents\Programming\C++\OpenGL_Projects\Cpp_OpenGL_SpinningCube\OpenGL_Shaders_Example\Spinning_Cube\Debug\GL_ShaderProgram.obj
 C:\Users\11307533\Documents\Programming\C++\OpenGL_Projects\Cpp_OpenGL_SpinningCube\OpenGL_Shaders_Example\Spinning_Cube\Spinning_Cube.cpp;C:\Users\11307533\Documents\Programming\C++\OpenGL_Projects\Cpp_OpenGL_SpinningCube\OpenGL_Shaders_Example\Spinning_Cube\Debug\Spinning_Cube.obj
