+++ OpenGL_Shaders_Example/Spinning_Cube/Debug/Spinning_Cube.tlog/link.secondary.1.tlog
@@ -1,2 +1,2 @@
-^C:\USERS\11307533\DOCUMENTS\PROGRAMMING\C++\OPENGL_PROJECTS\CPP_OPENGL_SPINNINGCUBE\OPENGL_SHADERS_EXAMPLE\SPINNING_CUBE\DEBUG\GL_ERRORHANDELING.OBJ|C:\USERS\11307533\DOCUMENTS\PROGRAMMING\C++\OPENGL_PROJECTS\CPP_OPENGL_SPINNINGCUBE\OPENGL_SHADERS_EXAMPLE\SPINNING_CUBE\DEBUG\GL_SHADERPROGRAM.OBJ|C:\USERS\11307533\DOCUMENTS\PROGRAMMING\C++\OPENGL_PROJECTS\CPP_OPENGL_SPINNINGCUBE\OPENGL_SHADERS_EXAMPLE\SPINNING_CUBE\DEBUG\SPINNING_CUBE.OBJ
+^C:\USERS\11307533\DOCUMENTS\PROGRAMMING\C++\OPENGL_PROJECTS\CPP_OPENGL_SPINNINGCUBE\OPENGL_SHADERS_EXAMPLE\SPINNING_CUBE\DEBUG\GL_DRAW.OBJ|C:\USERS\11307533\DOCUMENTS\PROGRAMMING\C++\OPENGL_PROJECTS\CPP_OPENGL_SPINNINGCUBE\OPENGL_SHADERS_EXAMPLE\SPINNING_CUBE\DEBUG\GL_ERRORHANDELING.OBJ|C:\USERS\11307533\DOCUMENTS\PROGRAMMING\C++\OPENGL_PROJECTS\CPP_OPENGL_SPINNINGCUBE\OPENGL_SHADERS_EXAMPLE\SPINNING_CUBE\DEBUG\GL_SHADERPROGRAM.OBJ|C:\USERS\11307533\DOCUMENTS\PROGRAMMING\C++\OPENGL_PROJECTS\CPP_OPENGL_SPINNINGCUBE\OPENGL_SHADERS_EXAMPLE\SPINNING_CUBE\DEBUG\SPINNING_CUBE.OBJ
 C:\Users\11307533\Documents\Programming\C++\OpenGL_Projects\Cpp_OpenGL_SpinningCube\OpenGL_Shaders_Example\Spinning_Cube\Debug\Spinning_Cube.ilk
