+++ OpenGL_Shaders_Example/Spinning_Cube/Spinning_Cube.vcxproj
@@ -133,11 +133,13 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
+    <ClCompile Include="GL_Draw.cpp" />
     <ClCompile Include="GL_ErrorHandeling.cpp" />
     <ClCompile Include="GL_ShaderProgram.cpp" />
     <ClCompile Include="Spinning_Cube.cpp" />
   </ItemGroup>
   <ItemGroup>
+    <ClInclude Include="GL_Draw.h" />
     <ClInclude Include="GL_ErrorHandeling.h" />
     <ClInclude Include="GL_ShaderProgram.h" />
     <ClInclude Include="GL_User_Types.h" />
