+++ OpenGL_Shaders_Example/Spinning_Cube/Spinning_Cube.vcxproj.filters
@@ -24,6 +24,9 @@
     <ClCompile Include="GL_ShaderProgram.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="GL_Draw.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="GL_ErrorHandeling.h">
@@ -33,6 +36,9 @@
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="GL_User_Types.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="GL_Draw.h">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
